@@ -1,26 +1,31 @@
 <?php
 
-
 namespace App\Http\Controllers\Api;
 
 
+use App\Events\SubscribedEvent;
+use App\Exceptions\PermissionDeniedException;
 use App\Http\Controllers\ApiController;
 use App\Http\Requests\CollectionRequest;
 use App\Models\Collection;
 use App\Repositories\CollectionRepository;
-use App\Traits\FollowTrait;
 use App\Transformers\CollectionTransformer;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 
 class CollectionController extends ApiController
 {
-    use FollowTrait;
-    protected static $model;
-
     public function __construct()
     {
-        static::$model = Collection::class;
+        $this->middleware('auth:api')->except(['show', 'index']);
+    }
+
+    // 订阅/取消订阅收藏集
+    public function toggleSubscribe(Collection $collection)
+    {
+        $result = me()->toggleSubscribe($collection);
+        if ($result['detached'] == []) {
+            event(new SubscribedEvent($collection, me()));
+        }
     }
 
     public function show(Collection $collection)
@@ -34,9 +39,20 @@ class CollectionController extends ApiController
         return $this->response()->noContent();
     }
 
+    /**
+     * @param Collection $collection
+     * @param CollectionRequest $request
+     * @param CollectionRepository $collectionRepository
+     * @return mixed
+     * @throws PermissionDeniedException
+     */
     public function update(Collection $collection, CollectionRequest $request, CollectionRepository $collectionRepository)
     {
+        if (!$collection->isAuthor()) {
+            throw new PermissionDeniedException('更新失败, 你的权限还不够喔 (╯︵╰,)');
+        }
         $collectionRepository->update($request->validated(), $collection);
+
         return $this->response()->noContent();
     }
 
@@ -44,21 +60,22 @@ class CollectionController extends ApiController
     {
         $collections = Collection::ApplyFilter($request)
             ->paginate($this->perPage());
-        return $this->response()->collection($collections, new CollectionTransformer());
+        return $this->response()->paginator($collections, new CollectionTransformer());
     }
 
     /**
      * @param Collection $collection
      * @return mixed
+     * @throws PermissionDeniedException
      * @throws \Exception
      */
     public function destroy(Collection $collection)
     {
-        if ($collection->user_id == auth()->id())
-            $collection->delete();
-        else {
-            throw new ModelNotFoundException('删除失败');
+        if (!$collection->isAuthor()) {
+            throw new PermissionDeniedException('删除失败, 你的权限还不够喔 (╯︵╰,)');
         }
+
+        $collection->delete();
         return $this->response()->noContent();
     }
 }
