@@ -4,7 +4,7 @@ use Faker\Generator as Faker;
 
 $factory->define(\App\Models\PostContent::class, function (Faker $faker) {
     return [
-        'post_id' => 1,
+        //'post_id' => random_int(1,5),
         'content' => $faker->text(200),
         'created_at' => \Carbon\Carbon::now(),
         'updated_at' => \Carbon\Carbon::now(),
