@@ -11,20 +11,24 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        factory(App\Models\User::class, 1)->create();
-        factory(App\Models\Post::class, 1)->create();
-        factory(App\Models\PostContent::class, 1)->create();
 
-        /*factory(App\Models\User::class, 5)->create()->each(function ($u) {
-            // 生成文章
-            $post = $u->posts()->save(factory(App\Models\Post::class)->make());
+        factory(App\Models\User::class, 5)->create();
+        factory(App\Models\Collection::class, 5)->create();
+        // 生成文章
+        factory(App\Models\Post::class, 5)->create()->each(function ($post) {
             // 生成文章正文
             $post->postContent()->save(factory(\App\Models\PostContent::class)->make());
-            // 生成文章点赞
-            //$u->likes()->create(['post_id' => $post->id]);
-        });*/
+        });
 
-        \App\Models\User::where('id', 1)->first()->update(['nickname' => 'admin', 'tel_num' => '15705547511', 'email' => '2013855675@qq.com']);
+        App\Models\User::where('id', 1)->first()->update([
+            'nickname' => 'admin',
+            'tel_num' => '15705547511',
+            'email' => '2013855675@qq.com'
+        ]);
 
+        // 清除缓存
+        Artisan::call('cache:clear');
     }
 }
+
+
